moved files into named folders on the client side so they are easier to find.
