+++ client/src/components/login/LoginForm.jsx
@@ -2,8 +2,8 @@
 import { useState } from 'react';
 import { Form, Button, Alert } from 'react-bootstrap';
 
-import { loginUser } from '../utils/API';
-import Auth from '../utils/auth';
+import { loginUser } from '../../utils/API';
+import Auth from '../../utils/authentication/auth';
 
 const LoginForm = () => {
   const [userFormData, setUserFormData] = useState({ email: '', password: '' });
