+++ client/src/components/login/SignupForm.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { Form, Button, Alert } from 'react-bootstrap';
 
-import { createUser } from '../utils/API';
-import Auth from '../utils/auth';
+import { createUser } from '../../utils/API';
+import Auth from '../../utils/authentication/auth';
 
 const SignupForm = () => {
   // set initial form state
