+++ client/src/components/navigation/Navbar.jsx
@@ -4,7 +4,7 @@ import { Navbar, Nav, Container, Modal, Tab } from 'react-bootstrap';
 import SignUpForm from './SignupForm';
 import LoginForm from './LoginForm';
 
-import Auth from '../utils/auth';
+import Auth from '../../utils/authentication/auth';
 
 const AppNavbar = () => {
   // set modal display state
