+++ client/src/pages/main-area/SavedBooks.jsx
@@ -7,9 +7,9 @@ import {
   Col
 } from 'react-bootstrap';
 
-import { getMe, deleteBook } from '../utils/API';
-import Auth from '../utils/auth';
-import { removeBookId } from '../utils/localStorage';
+import { getMe, deleteBook } from '../../utils/API';
+import Auth from '../../utils/authentication/auth';
+import { removeBookId } from '../../utils/storage/localStorage';
 
 const SavedBooks = () => {
   const [userData, setUserData] = useState({});
