+++ client/src/pages/main-area/SearchBooks.jsx
@@ -8,9 +8,9 @@ import {
   Row
 } from 'react-bootstrap';
 
-import Auth from '../utils/auth';
-import { saveBook, searchGoogleBooks } from '../utils/API';
-import { saveBookIds, getSavedBookIds } from '../utils/localStorage';
+import Auth from '../../utils/authentication/auth';
+import { saveBook, searchGoogleBooks } from '../../utils/API';
+import { saveBookIds, getSavedBookIds } from '../../utils/storage/localStorage';
 
 const SearchBooks = () => {
   // create state for holding returned google api data
