+++ client/src/utils/authentication/auth.js
